requirements.txt
requests===2.28.2
PyJWT===2.6.0